App cleaning when closing it fixed.
The code for cleaning the app stuff on the server side disappeared when the noactors code was started. Now,
when an app instance disappears, it removes the parent app node on ZK, which triggers the cleaning of the
app stuff on the server side.

This also makes to pass again the tests in TestSimpleNotification.