Use isSystemTable() in compactor check instead of enumeration of system tables.
NOTE: This change was added by Francis Liu as a pull request in the github repo. I've just re-sending it here to pass the Gerrit review. I'll remove the pull request in the github repo once this change is reviewed and merged.
Francis Liu wrote: Reviewing what the CP was doing and stumble on this small change. There are other system tables this forgets to take into account (ACLs and rsgroup) and we'd like to handle them in the same way as well.

Change-Id: I67913500641d55da2ac6103464e22778d04334f6