Make hbase client filters work with omid gets
Unfortunately you can't clone a get, and we we're copying over the
filter that the passed get contained. With the filter copied, they work
out of the box, as if the shadow cells are excluded by the filter they
will be copied when we check that they're in the snapshot. As column
prefix is the most likely filter to be used, we don't need to make this
any more efficient now as column prefix filters will also pull the
shadow cells also.

Change-Id: Ief5d2b6c0fb26ad847e94d5359a8b6ea51fbb775