Cleanup of logging
The log4j.xml in use expected yinst settings. This creates
$(LOG_FOLDER)out.log in the bin dir when running from an assembly built
tarball. This change changes the log4j.xml for the assembly to write to
omid.log in the main directory. The yinst log4j configuration is
unaffected.

It also adds a log4j configuration for tsobench that is separate from
the tso-server configuration and logs to console.

Change-Id: If5e6df1543ed0379f068049b8c5b1613408fd1e8