Economize bytes in suffix for shadow cells. Now it's a non-printable char
The shadow cell suffix is changed from :OMID_CTS to just a non-printable
char (ascii code 128)

WARNING: This is an backward incompatible change. Old clients will not
be able read data written by the new client, giving the possibility of
data inconsistency if a mix of old and new clients are running.

The combat this, I have added a handshake to the TSO server, to which
the capabilities available to omid are managed. To upgrade a cluster to
this change, first update the TSO. Old clients will no longer be able
connect to the TSO, so will not be able write data. New clients will
work fine with the TSO. New clients will also be able to read any data
previously written by the old clients, as it will check for both shadow
cell suffixes.

This change also adds checks to stop a client from accidentally writing
to shadow cells.

Change-Id: Ife89765f742799bd92a4918a2d0a0e59fd19942c