Drop tombstones during compaction
This change will drop tombstones, and any puts older than the tombstone
on compaction. However, the dropping only occurs if the tombstone is
older than the low watermark. Otherwise, a live transaction could try to
read a value older than the tombstone and find it missing, violating
snapshot isolation. This can still happen to really long running
transactions, whose start timestamp is lower than the low watermark, but
its not a problem because these won't be able to commit anyhow.

Change-Id: I5b1381130dfe6289857a0c523381da4fa18d1842