Fix dependency problem with ledgerset lib
We introduced this problem when adding HA to the commit-table cause
we used a snapshot version for ledgerset that required access to
the snapshot repository. However, when creating the client uber-jar
in the omid-hbase-client module, that repo was not accessible so
the snapshot ledgerset jar could not be downloaded.

In order to solve it, we updated the library to be a non-snapshot
one, so it can be reached from the standard release repository

Change-Id: I2c34c6429729f6b579c5bcaaf34a6bd33eaed446