Fix ledgerset dependencies problem with coprocessor
On an email on Feb 4 at 9:12 am Ganesan Sriram reported a problem when
deploying the coprocessor using version 2.2.21. When introducing HA in
version 2.2.19, the HBaseCommitTable included some source code that
involved references to the HBaseHACommitTable which in turn used
BookKeeper and Ledgerset.

The coprocessor just uses the HBaseCommitTable client for reading,
so it does not need the references to ledgerset and bookkeeper. To
solve the problem, the code that was the origin of those references
in HBaseCommitTable has been moved to a separate class.

Change-Id: I380372270f60b6a28a81ab1d8ea7bb275f376cdb