Improve isShadowCell() method to avoid data copies
The method signature has been changed to isShadowCell(Cell).
This avoids the data copies implied by the CellUtil.clone*() helper methods,
which are expensive, specially when they're called from the server side.

We avoid also to use the indexOf() method to check the shadow cell suffix. Now
we just check the last bytes of the qualifier in search for the shadow cell
suffix.

These facts relax the checking of qualifier names. However, from the user side,
we control the names of qualifiers in every Put and Delete operation by means
of the validateCell() method introduced in commit
1123158345a833c3937b19419fdd2403f91c9913 so this should not be a problem.

Change-Id: Id58a894c510b7416eab0290cb52930ffd05e3244