Allow async shadow cell update
HBaseTransactionManager has been modified to allow users select
sync or async shadow cells updates when creating the transaction
manager instance.

With async shadow cell updates, the control is returned to
the client earlier. It is also ensured that the transaction will be
removed from the commit table only after the shadow cells of the
transaction committed have been updated.

Change-Id: I0a75246d4be5b248096268ff8db904a3f499aef9