[SLINGCSCF-1689] Fix: Tables not flushed when updating shadow cells
Problem description: When autoflush is off and TTables are not closed after
transaction commit, a large number of cells without their corresponding
shadow cells may appear.

When the HBaseTransactionManager commits a TX, it:
1) writes the commit table entry for the Tx
2) updates shadow cells
3) erases the commit table entry

Before this fix, in step #2 the tables were not flushed, so some shadow cells
were not be written to HBase before the commit table entry was erased.

Change-Id: I8fceb0695ca8243b8f33e4ffa759ca33b7114757