[SLINGCSCF-1690] Remove legacy shadow cell suffix
NOTE: This change requires also to upgrade version to 2.3.x to
mark backward incompatible changes in very specific scenarios
(e.g. a new client with v2.3 can no longer talk to a TSO v2.2.x)

Having both suffixes may compromise performance. When a get or scan
operation is issued, for every qualifier requested TTable adds qualifiers
with both, the SHADOW_CELL_SUFFIX and the LEGACY_SHADOW_CELL_SUFFIX.
It also asks for the latest version of each column. This is fine for the
regular shadow cell suffix since we typically find it at the same version
as the requested qualifier. Unfortunately, for the legacy suffix HBase has
to spend time looking through every file (usually with its bloom filter,
but sometimes actually reading blocks) for a column that - in fact - does
not exist.

We still keep the handsake mechanism added in the commit that introduced
the legacy shadow cell suffix (1123158345a833c3937b19419fdd2403f91c9913)
because it can still be useful in the near future.

Change-Id: I406e98c7f2bbdc352d32ad4eb6693f54e3965c18