[SLINGCSCF-1959] Improve test coverage in commit-table module
Remove unused classes: DelayNullCommitTable
Move class InMemoryCommitTable used in tests to test sources

NOTE: We should also move NullCommitTable to test sources, but
it is used as the in-memory implementation of the TSOServer
when configured. So, we have two options to improve test
coverage in this module:

1) Remove the possibility of configuring the TSOServer with
an in-memory implementation if we determine that this is not
necessary anymore. This will clean-up also the config code
in TSOServer.

2) Add tests to NullCommitTable. This is the recomended
short term solution implemented in this commit.

Change-Id: I54e247f77df888486dd2845b9ad73126eaeb7095