[SLINGCSCF-2100] Fix coprocessor bug when clearing tombstones
The coprocessor removes the tombstone and all older versions
when the tombstone falls below the low watermark. This operation
is performed every time the co-processor runs on a minor or
major compaction.
However, this is not correct since during a minor compaction the
co-processor may only see a subset of store files and not have
the all the versions of a cell available for consideration.
Therefore, if it deletes a cell with a tombstone entry during a
minor compaction it may make an older version of the visible.

An "if" branch has been added in the coprocessor for removing
tombstones only in major compactions.

Also, a test for checking that tombstones are not removed during
minor compactions has been added.

Change-Id: Ifb0b807deda57889f73620c93668e637ee4b4b03