[SLINGCSCF-2230] Make compactor preserve cells deleted non-transactionally in tables (**)
(**) This commit allows the user to decide whether to preserve or not
the deleted cell on a minor compaction when the scanner is triggered

NOTE-A: No new tests have been added to check that non-transactional deleted
cells are removed upon a major compaction as this should happen in any
case.

NOTE-B: A test to check the user decission of retaining or not the
cell has been added.

This commit follows fixes of this problem in commits:

https://git.corp.yahoo.com/scalable-computing/omid/commit/b89c28a69e58fcee6993d6c756c4d5a5ed7530b5

and

https://git.corp.yahoo.com/scalable-computing/omid/commit/a1237820af1dcaacc29693dc5d2862f9850719af

Change-Id: I0814fe51fb3d49466aaa8bac1f017a6ab7dfeb6a