[SIEVE-127] TSO should fail fast when failures occur on creation
Since we introduced ZK as a first class component when launching the TSO,
if the specified network interface is not found, an illegal argument
exception is thrown. However, this error does not provoke the VM to
exit and the TSO gets stuck in the middle of nowhere.

Also this patch fixes the automatic assignment of the network interface
depending on the environment where we are (Mac or Linux)

Change-Id: Idd8f7b005d252903f4ee6ad9cd13a0e0a3556ce9