[SIEVE-74] Add epoch to RequestProcessor and allow reset its state
This is related to the TSOClient failover. When receiving client
requests from an old epoch it should return an abort to the client.

The epoch is moved from the TSOServer, where it was originally injected,
to the RequestProcessor which is the one that assigns it on creation or
reset, as is explained below.

This commit also allows to reset the state of the RequestProcessor.
A reset of the state of the RequestProcessor is required in the
launching of a new replica or when a backup replica becomes master.
A new method called resetState() has been added to the RequestProcessor
interface. The implementation basically consist of the same steps
that were doing in the previous constructor:

- Reseting the contents of the conflict resolution map
- Initialize the timestamp oracle to the last value stored
- Initialize the Low Watermark and persist it
- Initialize the epoch to the low watermark value

Tests have been fixed and some new added when required.

IMPORTANT NOTE: The epoch in the RequestProcessor is initialized but
never changes unless the state of the processor is reset. The check
for the HA implementation in the handleCommit() method originally
said: if (startTimestamp <= lowWatermark || startTimestamp <= epoch) {
However, in a replica that is master the new check added implies
that, at some point, all the transactions will be aborted. The
solution would be set the epoch value again when the value of the
low watermark changes but that means that the variable would be the
same as the low watermark which is also in the condition (redundant).
Also in a replica that has lost the lease it may occur that the
validation still succeeds.
To solve this we could:

1) inject here the leaseManager and ask if leaseManager.stillInLeasePeriod()
in order to check if this replica is legally allowed to commit the transaction:
if (startTimestamp <= lowWatermark || !leaseManager.stillInLeasePeriod()) {

2) In order to avoid complexity and spend unnecessary processor cycles in this
critical part of the system, and taking into account that we check for the
same condition in the PersistentProccessor before returning to the response
to the client, I would leave the original condition:
if (startTimestamp <= lowWatermark) {

After discussing it, we've chosen the solution described in 2)

Change-Id: Id8e7c546dd2883b90cadebfaccc3a51e11328cfc