[SIEVE-74] Extract LeaseManagement intf and move configuration to Guice
This change does not add any new functionality, just re-organizes the
implementation to provide more code flexibility and clarity in further
commits. A new interface to represent the required behavior for lease
management has been extracted from the previous LeaseManager
implementation. Then, two implementations are provided for that intf,
one intended to be used in HA and the other in scenarios involving
a single instance of the TSO without HA.

It also moves the configuration of the lease management to
a new Guice module that decides how to launch the lease management
instance (HA or Non-HA) based on the configuration parameter specified
for the TSO.

LeaseManager tests have been improved and a new test for the non-HA
lease manager has been added.

Change-Id: I5b4947e6840c6267838e07e86154da70bcc1822f