[SIEVE-74] Server lease check and reply to the client
The protobuf for the commit response message has been extended to
inform the client when is necessary to make heuristic actions to
determine the commit output.

The lease manager has been injected in the persistence processor to
check for master ownership and return the appropriate response to the
client (e.g. if heuristic actions need to be performed to determine
the commit output for the transaction.) New tests have been added
for the PersistenceProcessor implementation.

Also improved PersistenceProcessor's Batch tests.

Change-Id: Ia8e78530f4fb9ee3cc4f59304e5cb524615a7a42