[SIEVE-74] Refactor TSOServer initialization
This change does not add any new functionality, just re-organizes the
implementation to provide more code flexibility and clarity in further
commits.

In order to avoid dependency cycles with Guice, the state of the TSO
has been isolated and a TSOStateManager interface has been defined
for injecting it in those components that need to reset the TSO state.
Tests for that implementation have been added.

Also the initialization of the ZK client in the TSO server has been
improved to fix some potential issues that could occur when selecting
the ZK storage for the Timestamp Oracle and the HA mode at the same
time.

Change-Id: I5b2dfa30da07a38df8ae1b20b7a98fb5d26d89dc