[SIEVE-71] Add failover to TM
Added checks in the transaction manager (client side) to perform
heuristic actions during commit if the commit response received
requires that (e.g. because an old TSO re-appeared.) The heuristic actions
consists in trying to find the outcome of the Tx with the help of the
commit table, invalidating the transaction in case of undertainity.

TODO: At some point we'll need a mechanism to clean the CommitTable
from invalidated entries falling below the low watermark.

Change-Id: I1c91e5e18753a9683b84ae1a1fa1f9c07893ab7e