[SIEVE-490] Fix TTable.delete() method
Empty deleteP may occur in some scenarios throwing an
IllegalArgumentException, so we avoid that. Also a test
has been added to check that no write attempts to HBase
occur for example when a non-existing column family is
removed in a transactional delete.

Also other minor improvements:

1) Adjust ZK timeouts in client connections on the TSOClient and
TSOServer to avoid some tests failures when connecting
2) Adjust timeouts and fixes in tests classes

Change-Id: I8bce3d34eee1e0b0a9b3d4198bedadbcc0b0c235