Move Maven assembly descriptors to a "maven" dir in the root of each module
A specific directory named "maven/assembly" has been created in the root of each
project module requiring assembly in order to contain the configuration.
Before the assembly configuration was contained in the src/main
directory of each module.

Also the corresponding paths referring to the assembly conf location
have been modified in the poms of each module.

Change-Id: I82583de14fbe1d7149eaadba994ddb31757fb707