Fix tests for unconnected TSO clients
Ensure that after each request operation on an unconnected client
we return to the disconnected state in the client state machine.
Timing for assertion sheckings has been set to twice the reconnection
time set for the tests. This delay for checks must be enough to pass
tests in CI environment.

Change-Id: If31961855d55e329cbe8fff39921ff8738b45ec3