Compile tests related to TSOClient req&resp behavior in a single class
The goal was to reduce the time of execution of the previous individual tests.

The resulting class, TestTSOClientRequestAndResponseBehaviours,
encompasses tests related to TSOClient behaviour that were previously
included in the following classes:

- TestClientFailure [23 secs]
- TestClientTimeout [30 secs]
- TestDuplicateCommit [64 secs]
- TestHandshake [22 secs]
- TestTSOClientRetry [119 secs]
- TestTSOClientRetryPathOnTSOServer [64 secs]

The total execution time before the reorganization was  ~322 secs. Now
the execution time of the 13 tests included in the new class is ~44
secs.

Also, the test readability has been improved and all test use TestNG
assertions instead of the previous JUnit-to-TestNG wrapper.

Change-Id: Ia675794bbdf465ef0055e21ac0ee03546a6637af