Separate Omid-related config from HBase config
A new class OmidClientConfiguration has been created in the
transaction-client module in order to encompass all the
configuration related to Omid client (was included before in TSOClient)
This class is passed in the builder of the TSOClient

Similarly a new class -HBaseOmidClientConfiguration- has been created
to encompass the previous config plus the specific config options of Omid for
HBase (e.g. the commit table name). This class is passed in the builder
of the HBaseTransactionManager

This simplifies the configuration of the majority of tests and the user
examples. It also simplifies the wiring in the HBaseTransactionManager
builder and the creation of TSOClient instances.

NOTE: No functional changes have been added in this commit.

Remove builder pattern from config classes

Also:

- Metrics configuration param added to TSOClientConfiguration class

Change builder pattern for static factory to create TSOClient

Remove unnecesary parameters in AbstractTransactionManager

They were related to the closing of the passed resources to the constructor,
namely the TSOClient and the CommitTableClient intances

Change builder pattern for static factory to create HBaseTransactionManager

Also:
- Keep the builder pattern hidden for allow test configuration

Rename TSOClientConfiguration to OmidClientConfiguration

This is because the configuration on the client side can be used for
configuring not only the TSOClient but also the AbstractTransactionManager

Improve construction of HBaseTRansactionManager

Also:
- Remove unnecesary class OmidInstantiationException