Add infrastructure to create Omid website
The site is generated with the Apache site plugin and uses the
fluido skin, which is based on Twitter's bootstrap. The main pom file
contains its configuration and the site generation has been blocked
in the submodules, as they are not independent projects. All the site-related
configuration and documentation can be found in the "site" directory
and the generated static content to deploy that is generated with
mvn site:site and mvn site:staging is put to the "website" and "website-staging"
directories respectively.

When using mvn site:deploy, the website is directly pushed to the
Apache github repo by using the wagon-git extension (See wagon-git
extension in main pom.xml)

Change-Id: I55cb5bd9a0d1921433aec0ccdf2213f48221ebbe