[OMID-36] Use SystemExitPanicker for killing TSO process
Exit the JVM when an exception occurs in the TSO processors instead
of throwing a RuntimeException. With the Runtime exception, the process
hangs without exiting, which avoids the process to be relaunched in
both HA and non-HA scenarios.

Also terminate (close) all the TSO processor in an orderly manner when
shutdown.

Finally, log info has been added for both the startup and the shutdown
phases.

Change-Id: I0800f4fae98de29cbf02502db42700b0a1cdb280